Update last use factor source and send back failure from dialogs
